Rewrote threads from the ground
I decided to start at square one on threads when it became clear that
kthread did not include the features I needed to be able to process the
different scenarios I would encounter.  This should be better overall
and should be simpler as well.

It looks like you don't actually need to save registers everywhere
after all, since they're already in the interrupt stacks.  The solution
is to swap stacks, not registers.  This does mean changing the TSS stack
field on every context swap, which I had been against prior to this
change, but now I see that it is much simpler and overall better this
way.

Currently I have it just swapping from thread_zero to itself, which will
become the system idle task, and you can see the beginnings of this
where it schedules thread_zero if there are no runnable_threads.

Hopefully all continues as magically as this has and actually
supports my needs going forward.