pmm: groundwork for meory map
I want to do a physical meory map for a few reasons, mainly that I get
better visibility into memory utilization and it doesn't cost much
space.  This commit lays the groundwork for that system, but there are
still open questions about how to cound that data and wher to hook it
up.  See my comments at the bottom of pmm.c for more context on that.