Removed some task switch redundancies
Tasks now only switch if there is something to switch to
I don't just pop back and forth to an "idle thread" that
hlts in kernel_main anymore, which should be a big win overall.

Important note - now that locks work I have identified a very fatal
problem with the way I handle task switching.  Task switching currently
has a dependancy on allocating new memory with malloc.  Malloc
has a lock on it for heap consistency reasons.  These two facts
are fundamentally incompatible.  I rely on the other thread in a lock
contention releasing its lock so the blocked thread can continue,
but if the task switcher is blocked waiting for malloc to be available
it can never continue because it can never schedule anyone to run.

The task queue can still be of dynamic length, as long as any new
memory is allocated outside of the actual switch mechanism.  My
thought was to have a vector of runnable threads and an index into
that array representing the current thread.