threads: removed malloc from task switch
As noted in a recent commit, using locked resources like malloc in the
thread switch process is not good, since if the resource is taken and
task switch has to wait, nothing can ever release the lock.  This commit
removes that dependancy by reusing the memory explicitly.

I was already free()ing and then malloc()ing the same memory block in
practice, so it's fairly trivial to just literally reuse it.