syscalls: removed old do_syscall code
I was able to get strace functionality working using a table of format
strings fairly trivially, so I'm going ahead and ripping out the garbage
switch statement.

This may get harder in the future if I want to do more complicated
strace output, but I can handle that case in the future.