shell actually works over network
I had an epiphany that if I just added the abilty to swap out files on
file descriptors that I should be able to implement the shell on other
terminals (read: network sockets) pretty easily.

This commit adds sys_dup2, the syscall to do exactly that under POSIX,
and sure enough it took very little effort for me to be able to run
another copy of `init` using the network socket as stdin/stdout.

I call that a pretty major win.