I should probably actually set IDT entries
Not that they work, but baby steps
Also this contains a lot of fixes to printf in 32 bit mode to not
consume the wrong number of bytes.  There will be lots of buggy places
left in the code, but I'll find them in time.  I need to be certain that
specifically l and z mean 32/64 bit and ll means 64 bit.  Things that
work in x86_64 do not work in a calling convention where 64 bit values
have to be handled differently.

Also 300 commits to mainline woo