Remap kernel read-only pages read-only
This commit makes the kernel remap its code and rodata read-only at
early boot.  This caused me to find an interesting bug, in that it
appears the CPU edits the GDT when you do a segment move, and since the
GDT was mapped in rodata until this commit, 'mov ds, ebx' was failing in
return_from_interrupt.  The GDT is now mapped in .data and everything
appears to work.

What a good thing that I actually implemented kernel write protection!

In the future, I could just parse out the load directives in the
kernel's ELF header instead of relying on magic variables added at link
time, but this is good for the moment.