Added dmgr to handle resources identified by numbers
dmgr maps an integer (which could be extended to an arbitrary handle by
adding a mapping function) to a pointer.  It does not manage memory
besides that needed to store the pointers and a free list, so is very
small compared to the vecotrs it is replacing, and thus is replacing
large single allocations with no capability to reclaim holes with small
allocations and the ability to reuse identifiers once the old is no
longer needed.