Kernel threads work again
This was a tremendous problem.  Kernel threads would page fault on their
stack at start, then page fault again, causing a double fault and death.
This was because I was creating their pages as unbacked, and then not
poking them to cause the pages to have storage backing them.  The kernel
switches to the thread, it page faults because the stack doesn't exist,
but it can't push registers onto a stack that doesn't exist to be able
to handle the page fault.

With that corrected, I can keep working on the knetworker background
kernel task.