net: lots of pulling in the net system
Networking needs a lot of rethinking, pulled some things out of it that
I think are bad in preparation for some rethinking

Compiles right now, but networking definitely doesn't work.  The idea is
that I can start playing with the rtl driver independantly while I think
about how to move forward with that subsystem.