malloc: changed malloc so it can manage other pools
Basically just pulled the region_0 notion out into a parameter and
factored malloc/etc out into pool_malloc/etc(region_0).

The global allocator is unchanged except that I have brought back the
locking I had on the old malloc implementaion.  It is a client of the
new pool_* functions.

I intend to use this to actually manage mmap memory and for other
things.