shell stability improvements
This commit fixes a number of easy ways for the user to crash the shell.
This should improve the perceived polish of nightingale.

I found that the panic() I put to catch deadlocks was still there, thus
the OS would crash if any lock contention happened, I was printing a
character as a string in the history system, and I was making an attempt
to have the shell recover in case it crashed, but while (1) { execve }
is pretty dumb.

This also introduces the `crash` shell builtin, which does what it says
on the tin.  This is for testing.