threads: fixed thread switch deadlock
I removed the race condition where the process lock was checked at
thread switch and not taken, but did not consider the condition that I
was trying to avoid, the case where another thread tries to swap during
a swap.

This caused that condition to deadlock, as once a thread started to swap
out, if it did not have the process lock there would be no way to
resolve that contition.

This commit (re)introduces a very old mutex function, try_acquire_mutex,
which was removed in the time before I knew how to properly use C atomic
compare and swap.  This allows me to try to take the process lock
without blocking forever in a tack switch in the case that I don't have
it.  A cursory UI test did not yield any extant way for me to crash or
deadlock the UI and the test printf did fire several times.