test_runner: minimally working starting point
This commit adds a minimally working start for the test running system
to be used in Travis.

The test runner is an external Linux program that invokes `ruby run.rb`,
injects commands to stdin, and watches stdout for 'SUCCESS' or 'FAILURE'
or a crash.  It can handle running multiple tests, and it kills the
subprocesses after a set amount of time.