Major reorg.
I did some reorganization

This is pretty heavily inspired by the way thr Linux kernel is
organized, and puts more things at the top level.  Here's the new
theory:

kernel/ : the actual "kernel" or core of the OS.  Does not include data
structures or support libraries.

fs/ : anything related to filesystems or the virtual filesystem

ds/ : data structures, vector, queue, ringbuf, etc.

net/ : the network

drv/ : device drivers that are not core to functionality.  These are
good candidates to be made into kernel modules if I ever do that

arch/ : architecture specific things.

There is still work to do, arch/ is still pretty messy in my opinion and
I need a more consistent story around code being
architecture-independant.

The build process changes pretty significantly as well with this update,
each of the top level folders now builds into a statis library (e.g. the
kernel builds into libnightingale.a) which are then linked together in
the final build.

The idea is that these systems should be moderately independant as much
as possible.

This commit also moves lots of code out to a public include/ directory
hierarchy.  The vision is that local headers with no external use are
still in with the C files, but anything with exteral use is in the
global tree.  This should simplfy the list of folders that need to be
included from any given compilation.

Furthermore, there are still improvements left to be done in the user/
directory, but I think this is a good start and a good place to merge to
master.