New clone() syscall enabling multithreaded procs
This commit adds a new sys_clone() syscall that creates a new thread
within the same processes.  Currently this shares all of {virtual
memory, files}, unlike in Linux where you can choose what to clone and
what to keep.  The flags argument could be expanded to allow this in the
future.