Initial scaffold for loadable modules
This commit adds loadable kernel modules with the one tiny exception of
the part where I dind't realize I wouldn't be able to see the kernel ELF
headers directly and have to work out how they're deliverd through
multiboot.  Oh, and that 32bit isn't done.  But the linker seems to
work, so that's fun.

New SYS_LOADMOD syscall takes a file descriptor and loads the contents
into memory, performs relocations, and calls an `init_module` symbol
exported in the object.