Use lists for lots of things in thread.c
Lists of processes, lists of children, lists of threads

Improvements:
 - waitpid can scan for children needing an immidiate cleanup
 - you can print processes and their threads sensibly
 - you can get your process by pointer instead of a lookup

Threads and processes are no longer stored in malloc() manages memory,
they have dedicated object stores allocated by vmm_reserve, and a free
list is maintained of available slots in that region for housekeeping.

I imagine a world where those regions can be chained if the number of
processes or threads grows too large, but the amounts I chose for now
support ~1000 of each more or less.