Fix lots of problems with optimizations enabled
The x86_64 thread switch currently clobbers almost all of the x86_64
registers, (see new_thread_swap for the efforts to fix this), but this
was not being communicated to the compiler.  This meant anything
allocated to a register could be destroyed without warning on thread
swap at any time, and this was causing lots of greif when any
optimizations were enabled.  The i686 asm constraint appears to be
correct, which I am especially inclined to believe since I adaped that
code from an x86_32 OS.  There are still more problems with
optimizations enabled, but this fix makes a bunch of stuff work on -)g
and -O2 that did not before, including the `clone` exmaple program.

Lots of work yet to do, and preferably fixing the fact that we clobber
on swap in the longer term, but this is a good find.