Add fancy backtracing with symbols
Backtrace will now try to resolve the closest symbol to give an idea of
where the fault is.  It's not using DWARF debug information, so it can't
be as granular as addr2line, but it does show the function and offset of
the fault pretty well.

Still need to do a few things:
- resolve the symbol of the initial fault address and show that
- update ./dump.rb to accept the new format to still allow
line-debugging information.