Fix backtraces not working on crash
When on the static boot stack which was only 4096 bytes long, in the
case that I would crash inside printf, there would not be enough stack
space left to call printf again and display the backtrace.  The fix for
this is to expand the boot stack for the moment, with the expectation
that at some point in the future I should find away to detect and report
the stack smash condition.