CWD isn't fake anymore
I was thinking that I would try to fully emulate working directories in
userspace and just only use sys_xxat calls (openat, execveat, etc), but
this styarted leading to annoying consequences, so I changed it to an
entry on the thread (->cwd).  This seems pretty nice and actually seems
to work.