libc/errno: fix crash when errno is not a defined value
Previously calling perror() when errno was not between 0..ETODO would
have undefined behavior as it would overflow the error string array. I
found a crash related to this when I set errno to 2000 to find if a
codepath was being executed, so I want to be able to see errnos out of
range for this reason and other debugging. perror() now reports the
errno if it can't stringify it.