uart: let the system handle echoback of input
This is a first pass at moving the echo back load from my shell to the
system. I want to match the behavior seen in other systems where the
terminal automatically prints characters, even within library functions.
This is needed right now for the lua branch.

Lua does not perform character-buffered input in interactive mode.
Rather, it asks the system for a line with fgets(stdin). This means it's
not echoing the prompt characters back and I have no way to add that in
easily without changing that logic significantly.
That said, my shell assumes the opposite, so I can't support both
without even more legwork, so I'm choosing to go with the system-level
approach for now.
I had made some effort at doing this in the C library before I believe,
and that still mgiht be the correct path, nothing is certain.