syscalls: use signed type for syscall argument
I want to do this because it is more common that a syscall argument be
signed than unsigned. This should long term allow me to remove some
casts that are no longer necessary, but I will not attempt to do that
now.
This is not an ABI change, the binary representation is the same before
and after this change, but this makes supporting something like a
negative error number more natural, as sysret isn't unsigned/unsigned.

There's more work here in deciding how I actually want to represent
errors long-term - whether I like the { value, error flag } return or
not. This change makes no comment on a future for that decision.