tty: tty exists
I have learned why my old tty-ish system wasn't working particularly
well. I needed an extra step between the uart irq and the user program
to handle line buffering and interpret some input commands such as
control characters that are made in to signals, newline formats, and
also to handle echo back to the terminal device.

This is handled in the kernel because historically it improved latency
to not have to wake a user process on every keypress. I do want to
support this as an option, especially as it is important for the
functionality of much existing software (lua and sh come to mind
immidiately), while also supporting a more active mode of terminal
manipulation.

This will liekly involve some form of ioctl or similar mechanism, which
is something I will be looking at in the near future.

This change makes a number of other changes as well, it replaces
kernel/uart with kernel/serial. This would have been its own change and
does touch a number of annoyingly unrelated areas, the file system code
for example, but I made the changes concurrently and decided not to try
particularly hard to untabgle them.

Pseudoterminals are still an open question, but I think I understand the
problem space better after the realizations that let me to this point.