Migrate all headers to absolute paths
Write a little tool (migrate_paths.rb) to look for all the remaining
local paths from before the include/ migration and change them over to
absolute paths. Builds still work, so all appears good.

This is in preparation to stop requiring any -I arguments besides
include/ to the kernel build - local is fine but not if it's magically
local to a different directory.