build system: set CMake to cross compile
Without telling CMake that you're cross compiling, it will do things
like try to link a test program. This is no bueno when you're using a
cross compiler, since the test program will not link on the host system.
Setting CMAKE_SYSTEM_NAME indicates that you're performning a cross
compilation, and tells CMake not to sanity check things as it would
normally.