user: c++ support
This introduces simple and experimental support for C++ in usermode
programs. There is no C++ standard library. There are probably no global
object constructors called. This may not even stick around, but it's fun
to be able to say that it works and it introduces almost no additional
work since gcc can already compile C++.