simplify: remove the notion of "kernel modules"
This was the direction I was headed in already, that the kernel should
only be one blob of C (and asm) that gets built together, not some messy
muturally-recursive set of .a files that link together for no reason.

This commit removes the last edifice of the old "modules" system (though
loadable modules are totally still a thing). Building ngk from the outer
Makefile is now a simple

make -C kernel

And thus the world was improved.