c++ support: c++ in modules
This commit introduces support for kernel modules to be built in C++.
The same caveats apply as in the previous commit regarding usermode
support: There is no library, this is experimental, global contructors
are not called, and this may dissapear whenever I feel like it.

Still fun to see C++ code linked succesfully into the nightingale
kernel.