libm: fix building on I686
libm assumed that int32_t is an 'int' in a number of places, which
prevented it from building on I686, where i686-elf-gcc defines int32_t,
by way of the builtin macro __INT32_TYPE__, as 'long int'.

This change cleans up the places that were causing build failures and
builds on both X86_64 and I686.