thread: fix argv handling
After going on a big adventure, I found a few things wrong with argument
handling.

1. I was copying arguments out of userspace after loading the new
program image, thus anything static was trash.

2. I was not passing the program name as the first argument in argv in
some places inconsistently, which caused things like strace which are
opinionated about what is in what argument to behave badly.

3. I was (and continue to) copy arguments out of and in to the same argv
block in memory whenever I exec arguments. When strace runs execve(...,
argv + 1, ...), it's overwriting the existing argv with the new one in
place, which potentially introduces opportunities for things to go
wrong. So far it seems to work with the other problems fixed, but we'll
have to remember that it's a problem and keep an eye on it.