thread: cleanups and work toward signals
This commit changes some of the structures in the thread system, with a
few main overarching goals:

- make things more consistent and understandable
    (state/status -> thread_state/exit_status)
- make things more robust
    (THREAD_BLOCKED)
- add functionaly
    (blocked_list/node)

The main actual addition is that when a thread is blocked, it has a
pointer to the list and node it is blocked in, so it can be removed in
the event that it wakes up in some other means (signal cancels syscall,
thread is killed).

There's a bunch of things in here, and they are pretty intrusive
overall, so there are probably problems.