thread: new kill_thread / kill_process_group
In preparation for signals, this now lets you send messages to a
thread/process/group.

Most of the work for this is in previous commits, this is just the one
that fixed the fast that it didn't work for like 3 hours of me staring
at it trying to figure out what was happening.

The best commits.