tty: enable concurrent connections
The system can now accept and service serial lines COM1 and COM2,
providing a shell to each of them. This is made possible by generalizing
the previous notion of a tty to apply to each port.

This second line is available over a UNIX socket at ./serial2 when the
--serial2 option is provided to the run script.

This will be very useful for testing features that involve interprocess
communication, such as pipes and signals.