net: undelete net
This commit restores the networking subsystem to the nightingale system
gets it to a minimal standard of functionalty.

Currently, I can confirm the `net` program will do bidirectional UDP
communication with the linux host, and that's good enough to call this
okay for the moment.

There's a lot left to do of course, but this is good. Things working is
always good.

[TODO] fix all the things