malloc: use macro to poison pointers after free
I don't know if this will actually work long term, but it doesn't seem
to immidiately destroy everything, so let's keep it for now.

The idea is that per the C standard (C17/N2176):
6.2.4.2 Storage durations of objects
The value of a pointer becomes indeterminate when the object it points
to (or just past) reaches theend of its lifetime.

By this clause, the value of a pointer can change when the value it
describes is free'd, and therefore I take the liberty to change the
pointer to a trap representation (0d0d0d0d) that I will recognise as a
use-after-free. I should probably poison the actual memory as well.
Things to consider.