multiboot: map multiboot range before access
The kernel has reached such a size that accessing the multiboot
information structure can fall outside of the preallocated boot-time
memory map that I create for the kernel image. I haven't hit close
enough to that limit that I want to change it, but this way if GRUB
decides to load the multiboot image significatly far away from the rest
of the kernel (I have seen it up to several megabytes away), things
work.