dmgr: add locks
I think dmgr not having locks is one of the reasons I crash / drop
threads when there's a lot going on, so here's a trial implementation
that adds some.

There's a certain question of whether I should make this reentrant or
not. That could be accomplished with a depth counter that only actually
releases the mutex once it's reached 0 for a specific lock. That sounds
pretty easy, and would simplify some of the uses of mutexes, especially
malloc.