kernel timer: make generic timer_callback and do ticks
I want to be able to track some form of time, and the only way I know
how to do that is by having a ticking clock, so let's have a ticking
clock.

I also find that exposing myself to a constantly ticking clock exposes
more problems in the concurrency, so having that set should help harden
the system.