signals: fix lots of problems
Love these commits, tiny solution solves big problem.

This time, the problem was that signals were super random. They would
work sometimes, and other times the frame we were using was just trash.
Stuff like the kernel would try to execute the signal context on top of
its own stack, etc.

Turns out, this is because some large percentage of the interrupts that
come in are happening when the system is not in usermode. Those
interrupts still have a full frame dumped, except that that frame comes
from a kernel context.

This change makes it so the user_frame variable in the thread only gets
set if the frame is a usermode frame. So far this seems to significantly
cut down the amount of flakeyness in signals, and SEGV seems to work
pretty consistently now as well.

I'm mildly worried that we might have issues if the interrupt_frame gets
clobbered before it's used, so I may refactor this into a simple
last_rsp pointer in the frame, with no intent to save anything else.