signal: refactor stored froame pointer to user_sp
This commit makes good on the thought I had in the previous, since you
can't save the user frame every exception, it's not going to be
guaranteed that the pointer you have hasn't been clobbered. Therefore, I
now explicitly save the part I need from the frame and ignore the rest.

This commit also removes the disused interrupt_frame from the
signal_context in order to shrink the size of a thread struct.