fs: initial implementation of procfiles
This commit introduces the notion of a procfile to the nightingale
system. I don't know if I actually like the way this is implemented yet,
but having it available is better than not and I will continue to
iterate on the idea.

The key part here is that the text of the result is generated at open(2)
time and stored in the open_file structure itself, so that the text can
be read chunked without tearing or being regenerated, especially if
multiple processes call the procfile at the same time.

For the moment procfiles are read-only, but this could be changed easily
enough to support readonly or readwrite files.