x86/interrupt: add _with_ip to generic exception
I added this to page_fault long ago, but never added it to general
protection exceptions. This caused me to follow a bunch of rabbit holes
that I didn't need to as I was debugging the "dup2" file descriptor copy
problem earlier. I was missing the top function address, so I though the
faults were in `do_syscall_with_table`, and not `sys_write`.

There needs to be a general consistency pass on all of the exception
notification mechanisms in here, it's a right mess.