syscall: new strace formats
This commit shows [pid:tid] on strace outputs to show the change in
execution flow between threads. This would have been invaluable to have
when debugging the open file bug earlier, so hopefully I won't have to
go through that again.

This also changes numbers greater then 0x1000 to be hex and changes ->
ERRNO to -> <ERRNO> for some style points.