nc: revert pointer value poisonning
I'm no longer convinced this is a good idea to poison the actual pointer
value. I never caught any pointers in this form, I have lots of things
pointing to the same objects, and I use pointers to find objects in
datastructures.

Not to say it will never come back, but for now this is goodbye.