malloc: indicate uninitialized memory
This is silly to do unconditionally, especially since a lot of kernel
interfaces call `zmalloc` which 0-initializes the memory, so we set it
to 'M's and then set it right back again.

Hopefully this helps catches bugs - and especially indictaes the
difference between free'd memory and uninitialzied memory.

Once I started poisonning memory on free, I would then allocate it again
and think it was a free'd region when in fact it was just previously
freed memory.