user/init: don't start the second serial console
I want to rethink how I handle this console, so let's revert starting it
for now.

Init should probably have a while list of things it starts and not have
to make a bunch of tree intermediarys to do that.

Actually, maybe init should have *threads* that babysit children, so I
can flex that mucsle more!

I'll think about it.