shell: support pipes and redirection
This is a major update to the functionality of the nightingale shell,
adding support for unix pipes and file redirection.

This also updates some syscalls to bring them more in line with their
POSIX variants, as I was testing by porting the nightingale shell to
Linux.
Specifically, open() now accepts a third "mode" argument, and always
returns a writeable file descriptor if a new file is created.
Additionally, setpgid now accepts a pid and pgid to specify, though it
will not work for any other pid than the caller, and specifying an
arbitrary pgid has not been tested.

I know the parsing logic in the shell isn't necessarily 100%,
escpecially in error cases, but it seems to be close enough to working
that I can do interesting things that are syntactically valid. Further
robustness work is definitely needed.

This change also fixes sys_fork to increment refcounts on files, as my
pipes were being freed when they were close()d in the parent process.

I would have split this change up, but so much of this is
cross-dependant that I didn't want to take the time to untangle it (i.e.
create a working commit entirely in the old shell but with the pgid
change.)