thread: fix some use-after-frees
This is actually a pretty major reorg as these things go of the core
switch_thread functionality.

One of the main goals of this reorg is to make the differnt switch
reasons (SW_XXX) more clear, instead of sprinkling little exceptions all
over the switch code that you have to manually follow to understand
what's happening.

In the new world, switch thread starts out with a nice switch case block
with all the possible reasons, centralizing the information and making
things more clear.

I also implement a validation that a specific thread is not in the
runnable queue (assert_thread_is_not_runnable), which is placed in the
enqueue operations and before deallocating a thread. I do already have
the drop_thread function that would solve this problem equally well, but
it's important to me to catch and solve these problems rather than yolo
deleting things that may or may not be valid.

I'd like to give a big shoutout to the memory poisonning work from a
week or so ago, it continues to be the MVP with regards to accessing
freed pointers.