timer: centralize HZ in nc
There's a bigger question here of whether it's right to centralize these
kinds of kernel/nc shared data/datastructures in the nc headers.

It *is* a valid use case to use ng without nc (at runtime, not build
time), so I should not make it too onerous to get the header values that
the kernel essentially exports to userland as the source of truth
without combing through all of the headers for all of nc.

Maybe this is the real use for include/ng at toplevel? Make more things
follow syscall_consts?

The problem with that is that then building applications with nc depends
on the kernel headers. For now the tradeoff is that it's harder to port
a different userland entirely to ng, but easier to export nc to a
different kernel I suppose, and I suppose that is more in line with my
actual objectives as opposed to some philosphical goal about being able
to support an alternate userland on a system that barely supports one.

I'll have to keep thinking about it.