thread: new finalizer thread
This moves deinitialization of threads out of the threads own context
and to a kernel finalizer thread.

This is necessary because the timer can interrupt a thread after it has
freed its own context structure in do_thread_exit (in the previous
implementation). In that situation, the running thread is dead memory,
but a SW_TIMEOUT will use it. Being swapped out also means arbitrary
things woukd happen, including the reuse of the slot, etc.

I hope this is better than the previous situation, though I'm not
convinced there aren't still race conditions as implemented.