timer: stop running the same event multiple times
Previously, the timer was broken in the sense that if the timer action
was interrupted by a timer interrupt, the same timer action would be run
again.

This commit changes the structure of the timer callback to ensure that
the timer_head is popped before the function is run, ensuring this can't
happen.

I can't keep timer interrupts off until after the timer handler has run
because that would cause problems on thread switch with new threads,
where the timer handler gets jumped out of.