sys_create: don't create processes already running
This changes sys_create to leave the process not running initially,
allowing the programmer so call procstate(SETRUN) on it later once all
state has been prepared.

This also means that main() has to manually kick `init` off, which might
be better done elsewhere, but seems to work for now.