spalloc: new generalization of the special purpose allocator
spalloc is a general implementation of the "free slot" idiom I have used
in a number of places. Its intent is to centralize the work done on
those implementations and allow me to improve it specifically.

Mainly, I want to take the time to add region bumping to this, which
will remove the "maximum number" of items which I don't like.