kernel: crash process on illegal kernel action
This replaces the old behavior, panic().

This means the kernel doesn't crash if you sneeze on it nearly as much.
I do still need to replicate this across other kinds of faults (just
page faults for now).

This maybe should SIGSEGV?

This commit also adds a usermode program to test this behavior, crash.
This replaces the previous shell builtin.