pipe: fix problem where pipe would read EOF early
This commit fixes the problem where pipes would read EOF even if there
was still a process holding the other side open.

What was happening was that I was unconditionally setting signal_eof on
every close() call, but the creating processes close() both opposing ends
of the pipe right at the fork. This meant that unless the writing
process wrote data in before the reading process could pull the EOF, the
reader would assume there would never be data and move on.

This also updates process exit to close open file descriptors, as we did
the inverse for fork() already, so in order for refcounts to make any
sense at all, we need to close those files.