x86/uart: prevent irqs being lost under load
Previously if the CPU was under high load serial interrupts could get
lost and hang input for the user. This was happening because
write_to_serial_tty can do some pretty complicated things, including
potentially swap threads, which means irqs in progress can get dropped
and nevr EOI'd. This change should improve that significantly, by
readinf the value and then ACKing before writing to the TTY we increase
the risk of those events being processed slightly out of order in very
weird and extreme cases, but we don't hang for the user. I think that's
worth it.