c++: remove the effort to support c++ modules
Supporting C++ in the kernel is currently essentially impossible for a
few reasons. First of all, many of the headers assume C quite strongly,
to the point that the C++ modules have never built without turning
warnings down. This is an unideal solution and not one I want to
perpetuate. Secondly my inability to call global constructors hampers
the usefulness of C++ datstructures, for even those allocated at global
scope with 0s will not technically be initialized at access.

Half of the headers don't work at all anyway, so let's just be straight
up and say that this will not work. Maybe I will revisit this someday,
and userland is still totally good, but kernel C++ is just untenable at
this time.