thread: disallow reentrance to enqueue
This solves the deadlock problems where a thread gets dropped. It
happened because if you get preempted inside an enqueue, your queue will
be overwritten later. It's a data race on lists, so the quick fix is to
disallow reentrance. We could do some fancy lockfree business later, but
this will do for now.

There's a TODO here to make this work with SMP / with spinlocks too, but
that's a later problem I think.