multiboot: only map info on 64 bit.
This is still pretty fragile, since the bootloader can put the
information wherever it wants. That said, right now, the 32 bit one is
placed pretty close to the kernel, and the 64 bit one is placed higher.
This means if I remap the low one it sets my BSS to readonly on 32 bit.

This makes both 32 and 64 bit boot right this instant, but I'm actively
thinking about how we improve that.