user/sh: bugfix don't crash on full line command
Previously passing a full line comment to the shell would cause it to
try to look up a null file and crash the process in the kernel.

This patch adds logic to ensure something was written to the argument
buffer before executing anything.

The main motivation for this is to make shebang lines work.