elf: move ng/elf to linker submodule
This is an experiment with a new way to manage my dependancies. Instead
of maintaining two copies of the elf.c file, one in `linker`, and one
here, I will try managing it as a git submodule and making sure it
compiles both here and freestanding.

So far, it seems to be working, which is encouraging. Apparently git
submodules are annoying to manage, which I expect I will discover in
time.

I expect that this will be a good way to manage `ipstack` as well, and
potentially other things.