multiboot: map multiboot somewhere else
I'm tired of things breaking every time the kernel grows a tiny bit and
the bootloader decides to move the multiboot information.

This commit changes the mapping behavior a little for the multiboot
information, noting that the first 4 bytes of the multiboot info
structure have to be mapped already to get the length. After getting the
length, a new region is reserved and the physical memory associated with
multiboot is mapped there.

This seems to work.