rtl8139: actually support multiple NICs
Previously the initialization code fo the rtl was made to compile with
multiples of them, but there were major problems still.

This commit solves most of the problems -- NICs can recieve packets and
dispatch them properly including their net_if refrence for tracking.

This involves some sadness, like that NICs can occupy the same IRQ and I
have to handle that by going down the list every time there's an IRQ to
see if anyone has work to do.

This seems to work, just by sending packets to the interfaces and seeing
that the kernel doesn't like crash or anything, but I can't actually
test anything without a cleaner vision for addressing, routing, and
ipstack integration.